Refactor: extract shared mobile hooks to eliminate duplicated fetch logic
Extract usePriorMonthBalance, usePollsProjects, useHoaConfig from
OwnerDashboard3Cards, HOADashboard, and UnitSubDashboard into reusable
hooks. Removes ~120 lines of duplicated async fetch code that BugBot
flagged as a maintenance risk. Clean template for Admin mobile sprint.

Issue: #244
Made-with: Cursor
